Entity_Rel_Extractor/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-cors==0.8.1
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
art==6.2
asttokens==3.0.0
attrs==25.3.0
Authlib==1.6.0
beautifulsoup4==4.13.4
black==25.1.0
blinker==1.9.0
blis==0.7.11
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
cloudpickle==3.1.1
colorama==0.4.6
colorful==0.5.6
colorlog==6.9.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
datasets==2.14.7
debugpy==1.8.14
decorator==5.2.1
dill==0.3.7
distlib==0.3.9
docker-pycreds==0.4.0
docstring_parser==0.16
entrypoints==0.4
executing==2.2.0
faiss-cpu==1.8.0
Faker==37.4.0
fastapi==0.115.13
favicon==0.7.0
ffmpy==0.6.0
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2023.10.0
gitdb==4.0.12
GitPython==3.1.44
google-adk==1.4.1
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.97.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.21.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==4.37.2
gradio_client==1.0.2
graphviz==0.21
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.71.0
gunicorn==22.0.0
h11==0.16.0
htbuilder==0.9.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
identify==2.6.12
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.5.2
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.3.45
language_data==1.3.0
lightning==2.3.3
lightning-utilities==0.14.3
lxml==5.4.0
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mcp==1.9.4
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.5.0
multiprocess==0.70.15
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==1.43.1
neo4j==5.28.1
nest-asyncio==1.6.0
networkx==3.5
nodeenv==1.9.1
numpy==1.26.4
omegaconf==2.3.0
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.34.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
overrides==7.7.0
packaging==25.0
pandas==2.3.0
parso==0.8.4
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.8
pprintpp==0.4.0
pre_commit==4.2.0
preshed==3.0.10
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
py-spy==0.4.0
pyarrow==20.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pymdown-extensions==10.15
PyMuPDF==1.26.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytz==2025.2
pyvis==0.3.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
ray==2.35.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
ruff==0.12.0
safetensors==0.5.3
scikit-learn==1.5.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.30.0
setproctitle==1.3.6
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
sse-starlette==2.3.6
st-annotated-text==4.0.2
stack-data==0.6.3
starlette==0.46.2
streamlit==1.28.2
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.3.6
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.14.0
tbb==2021.13.1
tenacity==8.5.0
termcolor==2.4.0
thinc==8.2.5
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
torch==2.3.1
torchmetrics==1.7.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.41.2
typer==0.12.5
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.3
validators==0.35.0
vertexai==1.71.1
virtualenv==20.31.2
wandb==0.17.9
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
